refactored to work with new trays and calibration
