+++ aprs_vision/aprs_vision/slot_offsets.py
@@ -1,33 +1,32 @@
-from aprs_interfaces.msg import Object
-
+from aprs_interfaces.msg import Tray
 
 class SlotOffsets():
     offsets = {
-        Object.SMALL_GEAR_TRAY: {
-            'slot_1': (-0.028, 0.028),
-            'slot_2': (0.028, 0.028),
-            'slot_3': (-0.028, -0.028),
-            'slot_4': (0.028, -0.028),
+        Tray.SMALL_GEAR_TRAY: {
+            'slot_1': (-0.045, 0.045),
+            'slot_2': (0.045, 0.045),
+            'slot_3': (-0.045, -0.045),
+            'slot_4': (0.045, -0.045),
         },
-        Object.MEDIUM_GEAR_TRAY: {
-            'slot_1': (-0.038, 0.038),
-            'slot_2': (0.038, 0.038),
-            'slot_3': (-0.038, -0.038),
-            'slot_4': (0.038, -0.038),
+        Tray.MEDIUM_GEAR_TRAY: {
+            'slot_1': (-0.050, 0.050),
+            'slot_2': (0.050, 0.050),
+            'slot_3': (-0.050, -0.050),
+            'slot_4': (0.050, -0.050),
         },
-        Object.LARGE_GEAR_TRAY: {
-            'slot_1': (0.0, 0.055),
-            'slot_2': (0.0, -0.055),
+        Tray.LARGE_GEAR_TRAY: {
+            'slot_1': (-0.052, 0.06),
+            'slot_2': (0.052, 0.06),
         },
-        Object.M2L1_KIT_TRAY: {
-            'lg_1': (0.0, -0.04),
-            'mg_1': (-0.04, 0.055),
-            'mg_2': (0.04, 0.055),
+        Tray.M2L1_KIT_TRAY: {
+            'lg_1': (0.0, -0.075),
+            'mg_1': (-0.065, 0.0),
+            'mg_2': (0.065, 0.0),
         },
-        Object.S2L2_KIT_TRAY: {
-            'lg_1': (-0.055, 0.028),
-            'lg_2': (0.055, 0.028),
-            'sg_1': (-0.04, -0.055),
-            'sg_2': (0.04, -0.055),
+        Tray.S2L2_KIT_TRAY: {
+            'lg_1': (-0.052, 0.060),
+            'lg_2': (0.052, 0.060),
+            'sg_1': (-0.045, -0.045),
+            'sg_2': (0.045, -0.045),
         },
     }
