+++ aprs_vision/launch/aprs_vision.launch.py
@@ -1,94 +1,28 @@
 import os
-import math
+
+from ament_index_python.packages import get_package_share_directory
+
 from launch import LaunchDescription
+from launch_ros.actions import Node
+
 from launch.actions import (
-    DeclareLaunchArgument,
     OpaqueFunction,
 )
-from launch_ros.actions import Node
-from launch_ros.substitutions import FindPackageShare
-from launch.substitutions import PathJoinSubstitution, LaunchConfiguration
-from launch.conditions import IfCondition
-
 def launch_setup(context, *args, **kwargs):
 
-    fanuc_vision = Node(
+    parameters = os.path.join(get_package_share_directory('aprs_vision'), 'config', 'parameters.yaml')
+
+    fanuc_table_detection_area = Node(
         package='aprs_vision',
-        executable='fanuc_vision_node.py',
+        executable='detection_area_node.py',
+        name='table_detection_area',
+        namespace='fanuc',
+        parameters=[parameters],
         output='screen',
     )
 
-    teach_table_vision = Node(
-        package='aprs_vision',
-        executable='teach_table_vision_node.py',
-        output='screen'
-    )
-
-    motoman_vision = Node(
-        package='aprs_vision',
-        executable='motoman_vision_node.py',
-        output='screen'
-    )
-
-    motoman_conveyor_vision = Node(
-        package='aprs_vision',
-        executable='motoman_conveyor_node.py',
-        output='screen'
-    )
-
-    fanuc_conveyor_vision = Node(
-        package='aprs_vision',
-        executable='fanuc_conveyor_node.py',
-        output='screen'
-    )
-
-    fanuc_base_static_transform = Node(
-        package='tf2_ros',
-        executable='static_transform_publisher',
-        arguments=[
-            "--x" , f"{30 * .0254}", 
-            "--y" , f"{7 * .0254}",
-            "--frame-id" , "world",
-            "--child-frame-id" , "fanuc_base_link",
-        ]
-        # extarguments=[
-        #     {"FRAME_ID": 'world'},
-        #     {'CHILD_FRAME_ID': 'fanuc_base_link'},
-        #     {'X': 0},
-        #     {'Y': 0},
-        #     {'Z': 0},
-        # ]
-
-    )
-
-    motoman_base_static_transform = Node(
-        package='tf2_ros',
-        executable='static_transform_publisher',
-        arguments=[
-            "--x" , f"{77.5 * .0254}", 
-            "--y" , f"{6.5 * .0254}",
-            "--yaw" , f"{math.pi}",
-            "--frame-id" , "world",
-            "--child-frame-id" , "motoman_base_link",
-        ]
-        # arguments=[
-        #     {"FRAME_ID": 'world'},
-        #     {'CHILD_FRAME_ID': 'motoman_base_link'},
-        #     {'X': 0},
-        #     {'Y': 0},
-        #     {'Z': 0},
-        # ]
-
-    )
-    
     nodes_to_start = [
-        fanuc_vision,
-        teach_table_vision,
-        motoman_vision,
-        motoman_conveyor_vision,
-        fanuc_conveyor_vision,
-        # motoman_base_static_transform,
-        # fanuc_base_static_transform
+        fanuc_table_detection_area,
     ]
 
     return nodes_to_start
